Completely kill of all the explicit specialized classes in stnode
