+++ src/roman_datamodels/_stnode/__init__.py
@@ -6,7 +6,6 @@ The STNode classes and supporting objects generated dynamically at import time
 from __future__ import annotations
 
 from ._converters import *  # noqa: F403
-from ._explicit import *  # noqa: F403
 from ._node import *  # noqa: F403
 from ._schema import *  # noqa: F403
 from ._stnode import *  # noqa: F403
